feat: update log
