+++ src/handler/grpc/request/traces.rs
@@ -50,7 +50,24 @@ impl TraceService for TraceServer {
             return Err(Status::invalid_argument(msg));
         }
 
-        let request = ExportRequest::GrpcExportTraceServiceRequest(request);
+        let thread_id = metadata.get("thread_id");
+        let mut in_thread_id: usize = 0;
+        if let Some(thread_id) = thread_id {
+            in_thread_id = thread_id.to_str().unwrap().parse::<usize>().unwrap();
+        };
+
+        let stream_name = metadata.get(&CONFIG.grpc.stream_header_key);
+        let mut in_stream_name: Option<String> = None;
+        if let Some(stream_name) = stream_name {
+            in_stream_name = Some(stream_name.to_str().unwrap().to_string());
+        };
+
+        let request = ExportRequest::GrpcExportTraceServiceRequest((
+            org_id.unwrap().to_str().unwrap().to_string(),
+            in_thread_id,
+            request,
+            in_stream_name,
+        ));
         log::info!("begin to export flusher write");
         match self.flusher.write(request).await {
             Ok(resp) => match resp {
